:sparkles: feat: add validation in open file

+++ tests.sh
@@ -44,4 +44,15 @@ if [ $cr_cub -ne 1 ] || [ $cr_grep -ne 0 ]; then
     exit 1
 fi
 
+./cub3D ../maps/default.cub 2> stderr_output
+cr_cub=$?
+
+grep -q "Error: could not open the file" stderr_output
+cr_grep=$?
+
+if [ $cr_cub -ne 1 ] || [ $cr_grep -ne 0 ]; then
+    echo "Error in test 05: ./cub3D ../maps/default.cub"
+    exit 1
+fi
+
 rm stderr_output
